Updated code for better modularity
